+++ packages/mds-policy-service/tests/policy-client.spec.ts
@@ -1,6 +1,10 @@
 /* eslint-disable no-console */
-import { GeographyServiceClient, GeographyServiceManager } from '@mds-core/mds-geography-service'
-import { GeographyFactory, writePublishedGeography } from '@mds-core/mds-geography-service/tests/helpers'
+import {
+  GeographyFactory,
+  GeographyServiceClient,
+  GeographyServiceManager,
+  writePublishedGeography
+} from '@mds-core/mds-geography-service'
 import stream from '@mds-core/mds-stream'
 import { clone, days, now, START_ONE_MONTH_AGO, START_ONE_MONTH_FROM_NOW, uuid, yesterday } from '@mds-core/mds-utils'
 import { PolicyMetadataDomainModel } from '../@types'