+++ packages/mds-policy-service/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mds-core/mds-policy-service
 
+## 0.10.1
+
+### Patch Changes
+
+- d434c404: Import Geography helper methods from a real package, instead of a sub-module
+- Updated dependencies [d434c404]
+  - @mds-core/mds-geography-service@0.7.1
+
 ## 0.10.0
 
 ### Minor Changes